refactor so the optimize method can be used in other repo like dynamo
